refactor(options): move StyledCode into new shared/ subdirectory
